% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_integration.R
\name{create_oc_gt_table}
\alias{create_oc_gt_table}
\title{Create GT Table for Operating Characteristics}
\usage{
create_oc_gt_table(boinet_result, title = "Operating Characteristics")
}
\arguments{
\item{boinet_result}{Result object from boinet functions}

\item{title}{Optional table title}
}
\value{
A gt table object
}
\description{
Creates a formatted gt table displaying operating characteristics from
boinet simulation results.
}
