% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.iso.R
\name{multi.iso}
\alias{multi.iso}
\title{Model averaging of multiple unimodal isotopic regression}
\usage{
multi.iso(obs, n)
}
\arguments{
\item{obs}{Number of patients with events}

\item{n}{Number of patients}
}
\value{
Estimated probability
}
\description{
Given the location of the mode to be at each dose level, the unimodal
isotonically transformed values are calculated. A frequentist model
averaging approach is used to obtain the  estimated efficacy probability.
}
