% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boinet.R
\name{print.boinet}
\alias{print.boinet}
\title{Print boinet}
\usage{
\method{print}{boinet}(x, ...)
}
\arguments{
\item{x}{Object from boinet}

\item{...}{Other parameters}
}
\description{
boinet results are printed.
}
