% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gboinet.R
\name{print.gboinet}
\alias{print.gboinet}
\title{Print gboinet}
\usage{
\method{print}{gboinet}(x, ...)
}
\arguments{
\item{x}{Object from gboinet}

\item{...}{Other parameters}
}
\description{
gboinet results are printed.
}
