% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tite.boinet.R
\name{print.tite.boinet}
\alias{print.tite.boinet}
\title{Print tite.boinet}
\usage{
\method{print}{tite.boinet}(x, ...)
}
\arguments{
\item{x}{Object from \code{tite.boinet}.}

\item{...}{More options to pass to print.}
}
\value{
No return values. Key summary results from \code{tite.boinet} are displayed with
trial design settings.
}
\description{
Display key summary results from \code{tite.boinet}.
}
\seealso{
\code{\link{tite.boinet}}
}
