% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tite.gboinet.R
\name{print.tite.gboinet}
\alias{print.tite.gboinet}
\title{Print tite.gboinet}
\usage{
\method{print}{tite.gboinet}(x, ...)
}
\arguments{
\item{x}{Object from tite.gboinet}

\item{...}{Other parameters}
}
\description{
tite.gboinet results are printed.
}
