% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gboinet}
\alias{summary.gboinet}
\title{Summary method for gboinet objects}
\usage{
\method{summary}{gboinet}(object, ...)
}
\arguments{
\item{object}{A gboinet result object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Formatted summary output (invisible)
}
\description{
Summary method for gboinet objects
}
