% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tite.boinet}
\alias{summary.tite.boinet}
\title{Summary method for tite.boinet objects}
\usage{
\method{summary}{tite.boinet}(object, ...)
}
\arguments{
\item{object}{A tite.boinet result object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Formatted summary output (invisible)
}
\description{
Summary method for tite.boinet objects
}
