% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tite.gboinet}
\alias{summary.tite.gboinet}
\title{Summary method for tite.gboinet objects}
\usage{
\method{summary}{tite.gboinet}(object, ...)
}
\arguments{
\item{object}{A tite.gboinet result object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Formatted summary output (invisible)
}
\description{
Summary method for tite.gboinet objects
}
