% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.scoring.R
\name{utility.scoring}
\alias{utility.scoring}
\title{Utility defined by scoring}
\usage{
utility.scoring(probt, probe, psi00 = 40, psi11 = 60)
}
\arguments{
\item{probt}{Estimated toxicity probability}

\item{probe}{Estimated efficacy probability}

\item{psi00}{Score for toxicity=no and efficacy=no (default: 40)}

\item{psi11}{Score for toxicity=yes and efficacy=yes (default: 60)}
}
\value{
Utility value
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by scoring is calculated.
}
\examples{
utility.scoring(probt=c(0.1,0.2,0.4),probe=c(0.2,0.6,0.6));
}
