% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.truncated.linear.R
\name{utility.truncated.linear}
\alias{utility.truncated.linear}
\title{Utility defined by truncated linear function}
\usage{
utility.truncated.linear(
  probt,
  probe,
  tlow = 0.033,
  tupp = 0.462,
  elow = 0.21,
  eupp = 0.7
)
}
\arguments{
\item{probt}{Estimated toxicity probability}

\item{probe}{Estimated efficacy probability}

\item{tlow}{Lower bound for toxicity (default: 0.033)}

\item{tupp}{Upper bound for toxicity (default: 0.462)}

\item{elow}{Lower bound for efficacy (default: 0.21)}

\item{eupp}{Upper bound for efficacy (default: 0.70)}
}
\value{
Utility value
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by truncated linear functions is Calculated.
}
\examples{
utility.truncated.linear(probt=c(0.1,0.2,0.4),probe=c(0.2,0.6,0.6));
}
