% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.weighted.R
\name{utility.weighted}
\alias{utility.weighted}
\title{Utility defined by weighted function}
\usage{
utility.weighted(probt, probe, w1 = 0.33, w2 = 1.09, tox.upper = 0.462)
}
\arguments{
\item{probt}{Estimated toxicity probability}

\item{probe}{Estimated efficacy probability}

\item{w1}{Weight for toxicity-efficacy trade-off (default: 0.33)}

\item{w2}{Weight for penalty imposed on toxic doses (default: 1.09)}

\item{tox.upper}{Upper bound of toxicity probability (default: 0.462)}
}
\value{
Utility value
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by a weighted function is Calculated.
}
\examples{
utility.weighted(probt=c(0.1,0.2,0.4),probe=c(0.2,0.6,0.6));
}
