% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.logit.R
\name{fp.logit}
\alias{fp.logit}
\title{Fractional polynomial logistic regression}
\usage{
fp.logit(obs, n, dose)
}
\arguments{
\item{obs}{Number of patients with events.}

\item{n}{Number of patients.}

\item{dose}{Dose levels to be investigated.}
}
\value{
The \code{fp.logit} returns a vector of estimated probabilities for
each dose level.
}
\description{
Fractional polynomial (FP) logistic regression with two degrees of freedom is
performed to estimate the efficacy probabilities. The Best fitting FP model is
chosen by not taking into account the closed testing procedure.
}
