% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obd.select.R
\name{obd.select}
\alias{obd.select}
\title{Optimal biological dose selection}
\usage{
obd.select(
  probt = c(0.1, 0.2, 0.4),
  probe = c(0.2, 0.6, 0.6),
  method = c("utility.weighted", "utility.truncated.linear", "utility.scoring",
    "max.effprob"),
  phi = 0.33,
  phi1 = 0.033,
  phi2 = 0.462,
  delta = 0.7,
  delta1 = 0.42,
  tterm = c(0.5, 0.5, 0.5),
  eterm = c(0.5, 0.5, 0.5),
  stopT = 0.95,
  stopE = 0.95,
  w1 = 0.33,
  w2 = 1.09,
  tox.upper = 0.462,
  psi00 = 40,
  psi11 = 60
)
}
\arguments{
\item{probt}{Estimated toxicity probability (default: c(0.1,0.2,0.4))}

\item{probe}{Estimated efficacy probability (default: c(0.2,0.6,0.6))}

\item{method}{Method used for OBD selection}

\item{phi}{Target toxicity probability (default: 0.33)}

\item{phi1}{Lower bound of toxicity probability (default: 0.033)}

\item{phi2}{Upper bound of toxicity probability (default: 0.462)}

\item{delta}{Target efficacy probability (default: 0.70)}

\item{delta1}{Lower bound of efficacy probability (default: 0.42)}

\item{tterm}{Probability of meeting toxicity stopping criteria (default: c(0.5,0.5,0.5))}

\item{eterm}{Probability of meeting efficacy stopping criteria (default: c(0.5,0.5,0.5))}

\item{stopT}{Toxicity stopping criteria (default: 0.95)}

\item{stopE}{Efficacy stopping criteria (default: 0.95)}

\item{w1}{Weight for toxicity-efficacy trade-off (default: 0.33)}

\item{w2}{Weight for penalty imposed on toxic doses (default: 1.09)}

\item{tox.upper}{Upper bound of toxicity probability (default: 0.462)}

\item{psi00}{Score for toxicity=no and efficacy=no (default: 40)}

\item{psi11}{Score for toxicity=yes and efficacy=yes (default: 60)}
}
\value{
OBD
}
\description{
Optimal biological dose (OBD) is selected based on a method specified.
}
\examples{
obd.select(method="max.effprob");
}
