% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boinet.R
\name{print.boinet}
\alias{print.boinet}
\title{Print boinet}
\usage{
\method{print}{boinet}(x, ...)
}
\arguments{
\item{x}{Object from \code{boinet}.}

\item{...}{More options to pass to print.}
}
\value{
No return values. Key summary results from \code{boinet} are displayed with
trial design settings.
}
\description{
Display key summary results from \code{boinet}.
}
\seealso{
\code{\link{boinet}}
}
