% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gboinet.R
\name{print.gboinet}
\alias{print.gboinet}
\title{Print gboinet}
\usage{
\method{print}{gboinet}(x, ...)
}
\arguments{
\item{x}{Object from \code{gboinet}.}

\item{...}{More options to pass to print.}
}
\value{
No return values. Key summary results from \code{gboinet} are displayed with
trial design settings.
}
\description{
Display key summary results from \code{gboinet}.
}
\seealso{
\code{\link{gboinet}}
}
