% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tite.boinet.R
\name{print.tite.boinet}
\alias{print.tite.boinet}
\title{Print tite.boinet}
\usage{
\method{print}{tite.boinet}(x, ...)
}
\arguments{
\item{x}{Object from tite.boinet}

\item{...}{Other parameters}
}
\description{
tite.boinet results are printed.
}
