% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tite.gboinet.R
\name{print.tite.gboinet}
\alias{print.tite.gboinet}
\title{Print tite.gboinet}
\usage{
\method{print}{tite.gboinet}(x, ...)
}
\arguments{
\item{x}{Object from \code{tite.gboinet}.}

\item{...}{More options to pass to print.}
}
\value{
No return values. Key summary results from \code{tite.gboinet} are displayed with
trial design settings.
}
\description{
Display key summary results from \code{tite.gboinet}.
}
\seealso{
\code{\link{tite.gboinet}}
}
