% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tite.boinet.R
\name{tite.boinet}
\alias{tite.boinet}
\title{Simulation study of TITE-BOIN-ET design}
\usage{
tite.boinet(
  n.dose = 6,
  start.dose = 1,
  size.cohort = 3,
  n.cohort = 12,
  toxprob = c(0.01, 0.03, 0.06, 0.12, 0.18, 0.3),
  effprob = c(0.06, 0.08, 0.15, 0.25, 0.4, 0.8),
  phi = 0.33,
  phi1 = 0.033,
  phi2 = 0.462,
  delta = 0.7,
  delta1 = 0.42,
  alpha.T1 = 0.5,
  alpha.E1 = 0.5,
  tau.T = 30,
  tau.E = 45,
  accrual = 10,
  stopping.prob.T = 0.95,
  stopping.prob.E = 0.95,
  estpt.method = c("multi.iso", "fp.logistic", "obs.prob"),
  obd.method = c("utility.weighted", "utility.truncated.linear", "utility.scoring",
    "max.effprob"),
  w1 = 0.33,
  w2 = 1.09,
  psi00 = 40,
  psi11 = 60,
  n.sim = 1000,
  seed.sim = 66
)
}
\arguments{
\item{n.dose}{Number of dose (default: 6)}

\item{start.dose}{Starting dose (default: 1)}

\item{size.cohort}{Cohort size (default: 3)}

\item{n.cohort}{Number of cohort (default: 12)}

\item{toxprob}{True toxicity probability (default: c(0.01,0.03,0.06,0.12,0.18,0.30))}

\item{effprob}{True efficacy probability (default: c(0.06,0.08,0.15,0.25,0.40,0.80))}

\item{phi}{Target toxicity probability (default: 0.33)}

\item{phi1}{Lower bound of toxicity probability (default: 0.033)}

\item{phi2}{Upper bound of toxicity probability (default: 0.462)}

\item{delta}{Target efficacy probability (default: 0.70)}

\item{delta1}{Lower bound of efficacy probability (default: 0.42)}

\item{alpha.T1}{Probability that toxicity event occurs in the late half of assessment window (default: 0.5)}

\item{alpha.E1}{Probability that efficacy event occurs in the late half of assessment window (default: 0.5)}

\item{tau.T}{Assessment period for toxicity (days) (default: 30)}

\item{tau.E}{Assessment period for efficacy (days) (default: 45)}

\item{accrual}{Accrual rate (days) (default: 10)}

\item{stopping.prob.T}{Stopping probability for toxicity (default: 0.95)}

\item{stopping.prob.E}{Stopping probability for efficacy (default: 0.95)}

\item{estpt.method}{Methods to estimate efficacy probability}

\item{obd.method}{Methods to select OBD}

\item{w1}{Weight for toxicity-efficacy trade-off (default: 0.33)}

\item{w2}{Weight for penalty imposed on toxic doses (default: 1.09)}

\item{psi00}{Score for toxicity=no and efficacy=no (default: 40)}

\item{psi11}{Score for toxicity=yes and efficacy=yes (default: 60)}

\item{n.sim}{Number of simulated trial (default: 1000)}

\item{seed.sim}{Seed for random number generator (default: 66)}
}
\value{
Summary of simulation study results
}
\description{
Simulation study is conducted for TITE-BOIN-ET design.
}
\examples{
tite.boinet(estpt.method="obs.prob",obd.method="max.effprob");
}
