% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.scoring.R
\name{utility.scoring}
\alias{utility.scoring}
\title{Utility defined by scoring}
\usage{
utility.scoring(probt, probe, psi00, psi11)
}
\arguments{
\item{probt}{Estimated toxicity probability.}

\item{probe}{Estimated efficacy probability.}

\item{psi00}{Score for toxicity=no and efficacy=no.}

\item{psi11}{Score for toxicity=yes and efficacy=yes.}
}
\value{
The \code{utility.scoring} returns a utility value defined by the scoring.
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by scoring is calculated.
}
