% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.truncated.linear.R
\name{utility.truncated.linear}
\alias{utility.truncated.linear}
\title{Utility defined by truncated linear function}
\usage{
utility.truncated.linear(probt, probe, tlow, tupp, elow, eupp)
}
\arguments{
\item{probt}{Estimated toxicity probability}

\item{probe}{Estimated efficacy probability}

\item{tlow}{Lower threshold of toxicity linear truncated function.}

\item{tupp}{Upper threshold of toxicity linear truncated function.}

\item{elow}{Lower threshold of efficacy linear truncated function.}

\item{eupp}{Upper threshold of efficacy linear truncated function.}
}
\value{
Utility value
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by truncated linear functions is Calculated.
}
