% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.weighted.R
\name{utility.weighted}
\alias{utility.weighted}
\title{Utility defined by weighted function}
\usage{
utility.weighted(probt, probe, w1, w2, tox.upper)
}
\arguments{
\item{probt}{Estimated toxicity probability.}

\item{probe}{Estimated efficacy probability.}

\item{w1}{Weight for toxicity-efficacy trade-off.}

\item{w2}{Weight for penalty imposed on toxic doses.}

\item{tox.upper}{Upper bound of toxicity probability.}
}
\value{
Utility value
}
\description{
Given estimated toxicity and efficacy probabilities, the utility which is
defined by a weighted function is Calculated.
}
